// frontend/src/types/overtime-v2.ts
export enum StatusEnum {
  OPEN = 0,
  PAUSED = 1,
  RESOLVED = 10,
  CANCELLED = 255,
}

export enum StatusCodeEnum {
  OPEN = 'open',
  PAUSED = 'paused',
  RESOLVED = 'resolved',
  CANCELLED = 'cancelled',
  ONGOING = 'ongoing',
}

// Player Props
type PlayerProps = {
  playerId: number;
  playerName: string;
  playerScore?: number;
};

// Combined Position
export type CombinedPosition = {
  typeId: number;
  position: number;
  line: number;
};

// Odds
type Odds = {
  american: number;
  decimal: number;
  normalizedImplied: number;
};
// ...
export type Market = {
  gameId: string;
  sport: string;
  leagueId: number;
  leagueName: string;
  subLeagueId: number;
  typeId: number;
  type: string;
  line?: number;
  maturity: number;
  maturityDate: string;
  homeTeam: string;
  awayTeam: string;
  status: StatusEnum;
  isOpen: boolean;
  isResolved: boolean;
  isCancelled: boolean;
  isPaused: boolean;
  isOneSideMarket: boolean;
  isPlayerPropsMarket: boolean;
  isOneSidePlayerPropsMarket: boolean;
  isYesNoPlayerPropsMarket: boolean;
  playerProps: PlayerProps;
  combinedPositions: CombinedPosition[][];
  odds: Odds[];
  proof: string[];
  childMarkets?: Market[];
  statusCode: StatusCodeEnum;
  positionNames?: string[];
};